+++ packages/ui/src/main.tsx
@@ -37,7 +37,7 @@ Amplify.configure({
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
-        <RouterProvider router={router} />
+      <RouterProvider router={router} />
       <Toaster />
     </QueryClientProvider>
   </React.StrictMode>
