+++ packages/ui/src/routes/__root.tsx
@@ -1,24 +1,22 @@
-import React from "react";
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
 import type { QueryClient } from "@tanstack/react-query";
 
 const TanStackRouterDevtools =
   process.env.NODE_ENV === "production"
     ? () => null // Render nothing in production
-    : React.lazy(() =>
-        import("@tanstack/router-devtools").then((res) => ({
-          default: res.TanStackRouterDevtools,
-        }))
-      );
+    : () => null; // Render nothing in production
+// React.lazy(() =>
+//     import("@tanstack/router-devtools").then((res) => ({
+//       default: res.TanStackRouterDevtools,
+//     }))
+//   );
 
 export const Route = createRootRouteWithContext<{
   queryClient: QueryClient;
 }>()({
   component: () => (
     <>
-      <main className="">
-        <Outlet />
-      </main>
+      <Outlet />
       <TanStackRouterDevtools />
     </>
   ),
