+++ packages/ui/src/routes/auth/signin.tsx
@@ -2,7 +2,7 @@ import { createFileRoute, Link, useNavigate } from "@tanstack/react-router";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { Button } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/button";
 import {
   Form,
   FormControl,
@@ -14,6 +14,8 @@ import { Input } from "@/components/ui/input";
 import { useAuth } from "@/lib/hooks/use-auth";
 import { useToast } from "@/lib/hooks/use-toast";
 import { useEffect } from "react";
+import { cn } from "@/lib/utils";
+import { HomeIcon } from "lucide-react";
 
 export const Route = createFileRoute("/auth/signin")({
   component: RouteComponent,
@@ -50,9 +52,9 @@ function RouteComponent() {
   async function onSubmit(data: z.infer<typeof FormSchema>) {
     try {
       const signinStep = await auth.signIn(data.email, data.password);
-      
+
       if (signinStep === "DONE") {
-        navigate({ to: "/" });
+        navigate({ to: "/app" });
       } else {
         navigate({
           to: "/auth/confirm",
@@ -70,12 +72,22 @@ function RouteComponent() {
 
   useEffect(() => {
     if (auth.user) {
-      navigate({ to: "/" });
+      navigate({ to: "/app" });
     }
   }, [auth.user, navigate]);
 
   return (
     <div className="flex justify-center items-center min-h-screen">
+      <Link
+        to="/app"
+        className={cn(
+          buttonVariants({ variant: "ghost", size: "icon" }),
+          "absolute top-4 left-4"
+        )}
+      >
+        <HomeIcon className="w-6 h-6" />
+      </Link>
+
       <div className="w-[350px]">
         <Form {...form}>
           <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
